memmove ta dando segfault

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: mirkios <mirkios@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/12 10:41:19 by mde-figu          #+#    #+#             */
-/*   Updated: 2021/02/19 23:23:35 by mirkios          ###   ########.fr       */
+/*   Updated: 2021/02/19 23:26:01 by mirkios          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	unsigned long int	i;
+	size_t	i;
 	char			temp[n + 1];
 
 	i = 0;
